Reduce allocations in vectorized implementation
Cut down to 21.6M, 890MiB.
Mostly from Zygote by using MathConstants.e in the sigmoid function
instead of a float approximation (which was a workaround for a bug in
Zygote that I don't fully understand).
Also preallocated as much as I thought was reasonable and used in-place
and BLAS libraries for linear algebra.