Fix a thread-safety bug and change optimizations
The thread safety issue involved an in-place BLAS multiplication
that wrote to the same underlying destination memory. There's probably
not an easy way to multithread this while also avoiding allocations --
perhaps pre-allocate thread-local arrays outside the loop?

Also switch to an in-place map! instead of assignment, and use views
instead of slices in a couple of places.