Project structure changes
- Update package versions
- Rename to conform with Julia package conventions (hyphenated package
names didn't seem to be parsing properly in source code)
- Add direct dependencies to Project.toml to allow successful package
compilation
- Add test dependencies in Project.toml

- Include all files directly in the main src/AutomatonEdification.jl
file, instead of individually in the libraries or executables which use
them. This allows all library files to be visible to tests, and it
removes brittle include statements from most source files. Each file is
now loaded exactly once, with a clearly defined ordering.
- Fix using statements to keep executables working

- Add exports to most modules
- Distinguish between libraries (modules with exports, included in the
project) and executables

- Add a basic test/runtests.jl structure

- Update README with update instructions, new executable instructions,
and testing instructions